refactor: 퀴즈 업로드, 유닛 개요 업로드 중복 삭제 및 컴포넌트화
